+++ examples/localization.rs
@@ -4,14 +4,19 @@ use cushy::widget::MakeWidget;
 use cushy::{localize, Open, PendingApp};
 use unic_langid::LanguageIdentifier;
 
-fn localized() -> impl MakeWidget {
+fn localization() -> impl MakeWidget {
+    // Create a widget showing `message-hello-world`, which we will place on the
+    // window such that it detects the system locale.
     let element_in_default_locale = localize!("message-hello-world").contain();
 
+    // Create a widget showing `message-hello-world` in Spanish, always.
     let specific_locale: LanguageIdentifier = "es-ES".parse().unwrap();
     let elements_in_specific_locale = localize!("message-hello-world")
         .localized_in(specific_locale)
         .contain();
 
+    // Create a widget that shows `message-hello-world` in the locale selected
+    // by this example's available locales.
     let dynamic_locale: Dynamic<LanguageChoices> = Dynamic::default();
     let dynamic_message_label = localize!("message-hello-world");
 
@@ -31,33 +36,29 @@ fn localized() -> impl MakeWidget {
         .into_rows()
         .contain();
 
-    let bananas_counter = Dynamic::new(0u32);
+    // Fluent also supports parameterization, allowing localizers incredible
+    // flexibility in how messages and values are localized. This example shows
+    // how a dynamic counter can be used in localization in Cushy.
+    let bananas = Dynamic::new(0u32);
 
-    let counter_elements =
-        localize!("banana-counter-message", "bananas_counter" => &bananas_counter)
-            .and(
-                "+".into_button()
-                    .on_click(bananas_counter.with_clone(|counter| {
-                        move |_| {
-                            let mut counter = counter.lock();
-                            counter.checked_add(1).inspect(|new_counter| {
-                                *counter = *new_counter;
-                            });
-                        }
-                    })),
-            )
-            .and(
-                "-".into_button()
-                    .on_click(bananas_counter.with_clone(|counter| {
-                        move |_| {
-                            let mut counter = counter.lock();
-                            counter.checked_sub(1).inspect(|new_counter| {
-                                *counter = *new_counter;
-                            });
-                        }
-                    })),
-            )
-            .into_columns();
+    let counter_elements = localize!("banana-counter-message", "bananas" => &bananas)
+        .and("+".into_button().on_click(bananas.with_clone(|counter| {
+            move |_| {
+                let mut counter = counter.lock();
+                counter.checked_add(1).inspect(|new_counter| {
+                    *counter = *new_counter;
+                });
+            }
+        })))
+        .and("-".into_button().on_click(bananas.with_clone(|counter| {
+            move |_| {
+                let mut counter = counter.lock();
+                counter.checked_sub(1).inspect(|new_counter| {
+                    *counter = *new_counter;
+                });
+            }
+        })))
+        .into_columns();
 
     let dynamic_container = dynamic_message_label
         .and(counter_elements)
@@ -66,6 +67,7 @@ fn localized() -> impl MakeWidget {
         .contain()
         .localized_in(dynamic_locale.map_each(LanguageChoices::to_locale));
 
+    // Assemble the parts of the interface.
     element_in_default_locale
         .and(elements_in_specific_locale)
         .and(dynamic_container)
@@ -94,8 +96,12 @@ impl LanguageChoices {
 
 #[cushy::main]
 fn main(app: &mut PendingApp) -> cushy::Result {
-    // If you comment this block out, you can see the effect of having missing translation files.
+    // If you comment this block out, you can see the effect of having missing localization files.
     {
+        // Adds a localization for en-US, setting it as the default
+        // localization. If the system running this application is not
+        // compatible with the available locales, the `en-US` localization will
+        // be used.
         app.cushy().localizations().add_default(
             Localization::for_language(
                 "en-US",
@@ -103,6 +109,9 @@ fn main(app: &mut PendingApp) -> cushy::Result {
             )
             .expect("valid language id"),
         );
+        // Adds a localization for en-GB. Fluent supports region-specific
+        // localizations, and Cushy will attempt to find localizations in the
+        // best-matching locale.
         app.cushy().localizations().add(
             Localization::for_language(
                 "en-GB",
@@ -110,6 +119,7 @@ fn main(app: &mut PendingApp) -> cushy::Result {
             )
             .expect("valid language id"),
         );
+        // Adds a localization for es-ES.
         app.cushy().localizations().add(
             Localization::for_language(
                 "es-ES",
@@ -119,12 +129,12 @@ fn main(app: &mut PendingApp) -> cushy::Result {
         );
     }
 
-    localized().into_window().open(app)?;
+    localization().into_window().open(app)?;
 
     Ok(())
 }
 
 #[test]
 fn runs() {
-    cushy::example!(localized).untested_still_frame();
+    cushy::example!(localization).untested_still_frame();
 }