+++ src/lib.rs
@@ -397,6 +397,7 @@ fn initialize_tracing() {
             )
             .with(
                 Targets::new()
+                    .with_default(MAX_LEVEL)
                     .with_target("winit", Level::ERROR)
                     .with_target("wgpu", Level::ERROR)
                     .with_target("naga", Level::ERROR),